+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [Unreleased]
+### Fixed
+- Crash when loading enchantment token recipes if the enchantment has been removed [#340]
+
 ## [2.8.7] - 2018-10-17
 ### Added
 - Built-in recipes for enchantment tokens of Cyclic's enchantments [#340]